v0.4.1 Reorganise and rename functions
