+++ NAMESPACE
@@ -7,7 +7,6 @@ importFrom(methods, is)
 importFrom(R6, R6Class)
 importFrom(digest, digest)
 import(Matrix)
-exportPattern("^[[:alpha:]]+")
 S3method(print,mcml)
 S3method(summary,mcml)
 importFrom("utils", "capture.output")
