+++ R/R6covariance.R
@@ -134,7 +134,7 @@ Covariance <- R6::R6Class("Covariance",
                         #' @param parameters A vector of parameters for the covariance function(s). See Details.
                         update_parameters = function(parameters){
                           self$parameters <- parameters
-                          .Covariance__Update_parameters(private$ptr,parameters)
+                          Covariance__Update_parameters(private$ptr,parameters)
                           self$check(FALSE)
                         },
                         #' @description
@@ -146,7 +146,7 @@ Covariance <- R6::R6Class("Covariance",
                         #'                       parameters = c(0.05,0.8),
                         #'                       data= df)
                         print = function(){
-                          re <- .re_names(self$formula)
+                          re <- re_names(self$formula)
                           cat("\U2BC8 Covariance")
                           cat("\n   \U2BA1 Terms:",re)
                           cat("\n   \U2BA1 Parameters: ",self$parameters)
@@ -168,7 +168,7 @@ Covariance <- R6::R6Class("Covariance",
                         #' Returns the Cholesky decomposition of the covariance matrix D
                         #' @return A matrix
                         get_chol_D = function(){
-                          return(Matrix::Matrix(.Covariance__D_chol(private$ptr)))
+                          return(Matrix::Matrix(Covariance__D_chol(private$ptr)))
                         },
                         #' @description
                         #' The function returns the values of the multivariate Gaussian log likelihood
@@ -176,9 +176,9 @@ Covariance <- R6::R6Class("Covariance",
                         #' @param u Vector of random effects
                         #' @return Value of the log likelihood
                         log_likelihood = function(u){
-                          Q <- .Covariance__Q(private$ptr)
+                          Q <- Covariance__Q(private$ptr)
                           if(length(u)!=Q)stop("Vector not equal to number of random effects")
-                          loglik <- .Covariance__log_likelihood(private$ptr,u)
+                          loglik <- Covariance__log_likelihood(private$ptr,u)
                           return(loglik)
                         },
                         #' @description
@@ -186,7 +186,7 @@ Covariance <- R6::R6Class("Covariance",
                         #' with mean zero and covariance D.
                         #' @return A vector of random effect values
                         simulate_re = function(){
-                          re <- .Covariance__simulate_re(private$ptr)
+                          re <- Covariance__simulate_re(private$ptr)
                           return(re)
                         },
                         #' @description
@@ -196,9 +196,9 @@ Covariance <- R6::R6Class("Covariance",
                         #' @return None. Called for effects.
                         sparse = function(sparse = TRUE){
                           if(sparse){
-                            .Covariance__make_sparse(private$ptr)
+                            Covariance__make_sparse(private$ptr)
                           } else {
-                            .Covariance__make_dense(private$ptr)
+                            Covariance__make_dense(private$ptr)
                           }
                         },
                         #' @description
@@ -206,9 +206,9 @@ Covariance <- R6::R6Class("Covariance",
                         #' function term.
                         #' @return A data frame
                         parameter_table = function(){
-                          re <- .Covariance__re_terms(private$ptr)
-                          paridx <- .Covariance__parameter_fn_index(private$ptr)+1
-                          recount <- .Covariance__re_count(private$ptr)
+                          re <- Covariance__re_terms(private$ptr)
+                          paridx <- Covariance__parameter_fn_index(private$ptr)+1
+                          recount <- Covariance__re_count(private$ptr)
                           partable <- data.frame(id = paridx, term = re[paridx], parameter = self$parameters,count = recount[paridx])
                           return(partable)
                         }
@@ -223,26 +223,26 @@ Covariance <- R6::R6Class("Covariance",
                         cov_form = function(){
                           self$formula <- gsub("\\s","",self$formula)
                           self$formula <- gsub("~","",self$formula)
-                          re <- .re_names(self$formula)
+                          re <- re_names(self$formula)
                           self$formula <- re[1]
                           if(length(re)>1){
                             for(i in 2:length(re)){
                               self$formula <- paste0(self$formula,"+",re[i])
                             }
                           }
-                          private$ptr <- .Covariance__new(self$formula,
+                          private$ptr <- Covariance__new(self$formula,
                                                          as.matrix(self$data),
                                                          colnames(self$data))
-                          private$parcount <- .Covariance__n_cov_pars(private$ptr)
+                          private$parcount <- Covariance__n_cov_pars(private$ptr)
                           if(is.null(self$parameters))self$parameters <- rep(0.5,private$parcount)
-                          .Covariance__Update_parameters(private$ptr,self$parameters)
+                          Covariance__Update_parameters(private$ptr,self$parameters)
                           private$genD()
-                          self$Z <- .Covariance__Z(private$ptr)
+                          self$Z <- Covariance__Z(private$ptr)
                         },
                         genD = function(update=TRUE){
                           if(private$parcount != length(self$parameters))stop(paste0("Wrong number of parameters for covariance function(s). "))
-                          if(.Covariance__any_gr(private$ptr)).Covariance__make_sparse(private$ptr)
-                          D <- .Covariance__D(private$ptr)
+                          if(Covariance__any_gr(private$ptr))Covariance__make_sparse(private$ptr)
+                          D <- Covariance__D(private$ptr)
                           if(update){
                             self$D <- Matrix::Matrix(D)
                             private$hash <- private$hash_do()
