--- inst/include/glmmr/mhmcmc.ipp
@@ -1,5 +1,0 @@
-#ifndef MHMCMC_IPP
-#define MHMCMC_IPP
-
-
-#endif
