+++ src/other_functions.cpp
@@ -1,33 +1,18 @@
 #include <glmmr.h>
 
-// [[Rcpp::depends(RcppEigen)]]
+using namespace Rcpp;
 
-// [[Rcpp::export(.genX)]]
-Eigen::MatrixXd genX(const std::string& formula,
-                  const Eigen::ArrayXXd& data,
-                  const std::vector<std::string>& colnames) {
- glmmr::Formula form(formula);
- glmmr::LinearPredictor lin(form,data,colnames);
- return lin.X();
-}
-
-// // [[Rcpp::export(.x_names)]]
-// std::vector<std::string> x_names(const std::string& formula){
-//   glmmr::Formula form(formula);
-//   return form.fe_;
-// }
-
-// [[Rcpp::export(.re_names)]]
+// [[Rcpp::export]]
 std::vector<std::string> re_names(const std::string& formula){
- glmmr::Formula form(formula);
- std::vector<std::string> re(form.re_.size());
- for(int i = 0; i < form.re_.size(); i++){
-   re[i] = "("+form.z_[i]+"|"+form.re_[i]+")";
- }
- return re;
+  glmmr::Formula form(formula);
+  std::vector<std::string> re(form.re_.size());
+  for(int i = 0; i < form.re_.size(); i++){
+    re[i] = "("+form.z_[i]+"|"+form.re_[i]+")";
+  }
+  return re;
 }
 
-// [[Rcpp::export(.gen_dhdmu)]]
+// [[Rcpp::export]]
 Eigen::VectorXd gen_dhdmu(const Eigen::VectorXd& xb,
                           std::string family,
                           std::string link) {
@@ -36,7 +21,7 @@ Eigen::VectorXd gen_dhdmu(const Eigen::VectorXd& xb,
 }
 
 
-// [[Rcpp::export(.gen_sigma_approx)]]
+// [[Rcpp::export]]
 Eigen::MatrixXd gen_sigma_approx(const Eigen::VectorXd& xb,
                                  const Eigen::MatrixXd& Z,
                                  const Eigen::MatrixXd& D,
@@ -44,13 +29,13 @@ Eigen::MatrixXd gen_sigma_approx(const Eigen::VectorXd& xb,
                                  std::string link,
                                  double var_par,
                                  bool attenuate
-                                 ){
+){
   Eigen::MatrixXd S(xb.size(),xb.size());
   Eigen::VectorXd linpred(xb);
   if(attenuate){
     linpred = glmmr::maths::attenuted_xb(xb,Z,D,link);
   }
-
+  
   Eigen::VectorXd W = glmmr::maths::dhdmu(linpred,family,link);
   double nvar_par = 1.0;
   if(family=="gaussian"){
@@ -69,7 +54,7 @@ Eigen::MatrixXd gen_sigma_approx(const Eigen::VectorXd& xb,
   return S;
 }
 
-// [[Rcpp::export(.attenuate_xb)]]
+// [[Rcpp::export]]
 Eigen::VectorXd attenuate_xb(const Eigen::VectorXd& xb,
                              const Eigen::MatrixXd& Z,
                              const Eigen::MatrixXd& D,
@@ -78,9 +63,26 @@ Eigen::VectorXd attenuate_xb(const Eigen::VectorXd& xb,
   return linpred;
 }
 
-// [[Rcpp::export(.dlinkdeta)]]
+// [[Rcpp::export]]
 Eigen::VectorXd dlinkdeta(const Eigen::VectorXd& xb,
-                            const std::string& link){
- Eigen::VectorXd deta = glmmr::maths::detadmu(xb,link);
- return deta;
+                          const std::string& link){
+  Eigen::VectorXd deta = glmmr::maths::detadmu(xb,link);
+  return deta;
 }
+
+// This is a function in development - it works as expected,
+// but is subject to current research and so it is currently
+// not exposed to the user through the model class yet
+// [[Rcpp::export]]
+SEXP girling_algorithm(SEXP xp, SEXP N_,
+                       SEXP sigma_sq_, SEXP C_,
+                       SEXP tol_){
+  double N = as<double>(N_);
+  double sigma_sq = as<double>(sigma_sq_);
+  double tol = as<double>(tol_);
+  Eigen::VectorXd C = as<Eigen::VectorXd>(C_);
+  XPtr<glmmr::Model> ptr(xp);
+  Eigen::ArrayXd w = ptr->optimum_weights(N,sigma_sq,C,tol);
+  return wrap(w);
+}
+
